events refactor
